requirements.txt
telebot
mysql-connector-python
requests
beautifulsoup4
